Ai_model/pyproject.toml
[project]
name = "ai-model"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.11.0",
    "fastapi>=0.115.12",
    "google-generativeai>=0.8.5",
    "grpcio>=1.72.1",
    "grpcio-tools>=1.71.0",
    "langchain>=0.3.25",
    "langchain-google-genai>=2.0.10",
    "langchain-pinecone>=0.2.8",
    "numpy>=2.2.6",
    "pdf2image>=1.17.0",
    "pika>=1.3.2",
    "pypdf2>=3.0.1",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "redis>=6.2.0",
    "shiv>=1.0.8",
    "uvicorn>=0.34.3",
]
